exercise 3

+++ things/models.py
@@ -5,4 +5,4 @@ from  django.core.validators import MinValueValidator, MaxValueValidator
 class Thing(models.Model):
     name = models.CharField(max_length=30, unique=True, blank=False)
     description = models.TextField(blank=True, max_length=120, unique=False)
-    quantity = models.PositiveSmallIntegerField(default=0)
+    quantity = models.IntegerField(default=0)
